// sdks/remote_plugin/nodejs/src/proto/plugin.ts
export enum RuntimeMessageType {
  RUNTIME_MESSAGE_TYPE_UNSPECIFIED = 0,
  RUNTIME_MESSAGE_TYPE_LIFECYCLE = 1,
  RUNTIME_MESSAGE_TYPE_SIGNAL = 2,
  RUNTIME_MESSAGE_TYPE_PAYLOAD = 3,
  RUNTIME_MESSAGE_TYPE_EMIT_SIGNAL = 4,
  RUNTIME_MESSAGE_TYPE_EMIT_PAYLOAD = 5,
  RUNTIME_MESSAGE_TYPE_EMIT_LOG = 6,
  RUNTIME_MESSAGE_TYPE_ACK = 7,
  RUNTIME_MESSAGE_TYPE_REPORT = 8,
  RUNTIME_MESSAGE_TYPE_CANCEL = 9,
  UNRECOGNIZED = -1,
}
// ...
export enum LifecycleType {
  LIFECYCLE_TYPE_UNSPECIFIED = 0,
  LIFECYCLE_TYPE_START = 1,
  LIFECYCLE_TYPE_READY = 2,
  LIFECYCLE_TYPE_PAUSE = 3,
  LIFECYCLE_TYPE_RESUME = 4,
  LIFECYCLE_TYPE_STOP = 5,
  UNRECOGNIZED = -1,
}
// ...
export enum LogLevel {
  LOG_LEVEL_UNSPECIFIED = 0,
  LOG_LEVEL_DEBUG = 1,
  LOG_LEVEL_INFO = 2,
  LOG_LEVEL_WARN = 3,
  LOG_LEVEL_ERROR = 4,
  UNRECOGNIZED = -1,
}
// ...
export interface RuntimeMessage {
  instanceId: string;
  messageId: string;
  correlationId: string;
  type: RuntimeMessageType;
  metadata: { [key: string]: string };
  body:
    | { $case: "lifecycle"; lifecycle: LifecycleEvent }
    | { $case: "signal"; signal: SignalEvent }
    | { $case: "payload"; payload: PayloadEvent }
    | { $case: "emitSignal"; emitSignal: EmitSignal }
    | { $case: "emitPayload"; emitPayload: EmitPayload }
    | { $case: "emitLog"; emitLog: EmitLog }
    | { $case: "report"; report: EventReport }
    | { $case: "ack"; ack: EventAck }
    | { $case: "cancel"; cancel: CancelEvent }
    | undefined;
}

export interface RuntimeMessage_MetadataEntry {
  key: string;
  value: string;
}

export interface LifecycleEvent {
  type: LifecycleType;
}

export interface SignalEvent {
  name: string;
  properties: Buffer;
}

export interface PayloadEvent {
  name: string;
  properties: Buffer;
}

export interface EmitSignal {
  signal: SignalEvent | undefined;
  port: number;
}

export interface EmitPayload {
  payload: PayloadEvent | undefined;
  port: number;
}

export interface EmitLog {
  level: LogLevel;
  message: string;
  fields: { [key: string]: string };
}

export interface EmitLog_FieldsEntry {
  key: string;
  value: string;
}

export interface EventReport {
  status: ReportStatus;
  error: RemoteError | undefined;
}

export interface RemoteError {
  code: string;
  message: string;
  details: string;
}

export interface EventAck {
  timestamp: number;
}

export interface CancelEvent {
}
// ...
export const LifecycleEvent: MessageFns<LifecycleEvent> = {
  encode(message: LifecycleEvent, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== 0) {
      writer.uint32(8).int32(message.type);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LifecycleEvent {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLifecycleEvent();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): LifecycleEvent {
    return { type: isSet(object.type) ? lifecycleTypeFromJSON(object.type) : 0 };
  },

  toJSON(message: LifecycleEvent): unknown {
    const obj: any = {};
    if (message.type !== 0) {
      obj.type = lifecycleTypeToJSON(message.type);
    }
    return obj;
  },

  create(base?: DeepPartial<LifecycleEvent>): LifecycleEvent {
    return LifecycleEvent.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<LifecycleEvent>): LifecycleEvent {
    const message = createBaseLifecycleEvent();
    message.type = object.type ?? 0;
    return message;
  },
};

function createBaseSignalEvent(): SignalEvent {
  return { name: "", properties: Buffer.alloc(0) };
}

export const SignalEvent: MessageFns<SignalEvent> = {
  encode(message: SignalEvent, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    if (message.properties.length !== 0) {
      writer.uint32(18).bytes(message.properties);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SignalEvent {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSignalEvent();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.properties = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SignalEvent {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      properties: isSet(object.properties) ? Buffer.from(bytesFromBase64(object.properties)) : Buffer.alloc(0),
    };
  },

  toJSON(message: SignalEvent): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.properties.length !== 0) {
      obj.properties = base64FromBytes(message.properties);
    }
    return obj;
  },

  create(base?: DeepPartial<SignalEvent>): SignalEvent {
    return SignalEvent.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<SignalEvent>): SignalEvent {
    const message = createBaseSignalEvent();
    message.name = object.name ?? "";
    message.properties = object.properties ?? Buffer.alloc(0);
    return message;
  },
};

function createBasePayloadEvent(): PayloadEvent {
  return { name: "", properties: Buffer.alloc(0) };
}

export const PayloadEvent: MessageFns<PayloadEvent> = {
  encode(message: PayloadEvent, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    if (message.properties.length !== 0) {
      writer.uint32(18).bytes(message.properties);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PayloadEvent {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePayloadEvent();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.properties = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PayloadEvent {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      properties: isSet(object.properties) ? Buffer.from(bytesFromBase64(object.properties)) : Buffer.alloc(0),
    };
  },

  toJSON(message: PayloadEvent): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.properties.length !== 0) {
      obj.properties = base64FromBytes(message.properties);
    }
    return obj;
  },

  create(base?: DeepPartial<PayloadEvent>): PayloadEvent {
    return PayloadEvent.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<PayloadEvent>): PayloadEvent {
    const message = createBasePayloadEvent();
    message.name = object.name ?? "";
    message.properties = object.properties ?? Buffer.alloc(0);
    return message;
  },
};

function createBaseEmitSignal(): EmitSignal {
  return { signal: undefined, port: 0 };
}

export const EmitSignal: MessageFns<EmitSignal> = {
  encode(message: EmitSignal, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.signal !== undefined) {
      SignalEvent.encode(message.signal, writer.uint32(10).fork()).join();
    }
    if (message.port !== 0) {
      writer.uint32(16).int32(message.port);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): EmitSignal {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEmitSignal();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.signal = SignalEvent.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.port = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): EmitSignal {
    return {
      signal: isSet(object.signal) ? SignalEvent.fromJSON(object.signal) : undefined,
      port: isSet(object.port) ? globalThis.Number(object.port) : 0,
    };
  },

  toJSON(message: EmitSignal): unknown {
    const obj: any = {};
    if (message.signal !== undefined) {
      obj.signal = SignalEvent.toJSON(message.signal);
    }
    if (message.port !== 0) {
      obj.port = Math.round(message.port);
    }
    return obj;
  },

  create(base?: DeepPartial<EmitSignal>): EmitSignal {
    return EmitSignal.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<EmitSignal>): EmitSignal {
    const message = createBaseEmitSignal();
    message.signal = (object.signal !== undefined && object.signal !== null)
      ? SignalEvent.fromPartial(object.signal)
      : undefined;
    message.port = object.port ?? 0;
    return message;
  },
};

function createBaseEmitPayload(): EmitPayload {
  return { payload: undefined, port: 0 };
}

export const EmitPayload: MessageFns<EmitPayload> = {
  encode(message: EmitPayload, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.payload !== undefined) {
      PayloadEvent.encode(message.payload, writer.uint32(10).fork()).join();
    }
    if (message.port !== 0) {
      writer.uint32(16).int32(message.port);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): EmitPayload {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEmitPayload();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.payload = PayloadEvent.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.port = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): EmitPayload {
    return {
      payload: isSet(object.payload) ? PayloadEvent.fromJSON(object.payload) : undefined,
      port: isSet(object.port) ? globalThis.Number(object.port) : 0,
    };
  },

  toJSON(message: EmitPayload): unknown {
    const obj: any = {};
    if (message.payload !== undefined) {
      obj.payload = PayloadEvent.toJSON(message.payload);
    }
    if (message.port !== 0) {
      obj.port = Math.round(message.port);
    }
    return obj;
  },

  create(base?: DeepPartial<EmitPayload>): EmitPayload {
    return EmitPayload.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<EmitPayload>): EmitPayload {
    const message = createBaseEmitPayload();
    message.payload = (object.payload !== undefined && object.payload !== null)
      ? PayloadEvent.fromPartial(object.payload)
      : undefined;
    message.port = object.port ?? 0;
    return message;
  },
};